// src/types.ts
export interface Usuario {
  uid: string;
  nome: string;
  email: string;
  empresaId: string;
  unidadeId?: string;
  papel: 'admin' | 'supervisor' | 'operador' | string;
  cargo?: string;
  matricula?: string;
  modulosPermitidos?: string[];
  status: 'ativo' | 'inativo';
  criadoEm?: any;
  mfaHabilitado?: boolean;
  mfaSegredo?: string;
  isControle?: boolean;
}
// ...
export interface ValidadeRow {
  _docId?: string;
  id?: number | string;
  empresaId?: string;
  codigo: string;
  descricao: string;
  palhete: number;
  lastro: number;
  caixa: number;
  quantidade?: number;
  validade: string;
  diasParaVencer?: number;
  localizacao: 'central' | 'pnc' | 'repack' | 'picking' | 'marketplace' | string;
  bloco?: string;
  lote?: string;
  responsavel?: string;
  dataColeta?: string;
  semanaNumero?: number;
  mesReferencia?: string;
  dataTransferenciaPnc?: string;
  totalUnities?: number;
  totalUnitiesRaw?: number;
  valorTotal?: number;
  hlTotal?: number;
  vidaUtil?: number;
  cadastradoPor?: string;
  origem?: string;
  cadastradoEm?: string;
  recontadoEm?: string;
  atualizadoEm?: string;
  restauradoEm?: string;
  atualizadoPor?: string;
  _criadoEm?: string;
}